Moved historyEntries file to component folder
